Time: 5 ms (51.92%), Space: 6 MB (70.58%) - LeetHub

+++ 0074-search-a-2d-matrix/0074-search-a-2d-matrix.c
@@ -1,5 +1,5 @@
 bool searchMatrix(int** matrix, int matrixSize, int* matrixColSize, int target) {
-    short int   med = 0;
+    int8_t      med = 0;
     uint8_t     row = 0;
     uint8_t     left = 0;
